Implemented neuron class

+++ src/main/java/com/pat_eichler/bnn/brain/GeneticsModel.java
@@ -12,7 +12,7 @@ public abstract class GeneticsModel extends Genetics{
         super(dna);
     }
 
-    public abstract byte getNeuronStateChange(byte[] preNeuronStateCounts, byte[] postNeuronStateCounts, byte curState);
+    public abstract byte getNeuronStateChange(short[] preNeuronStateCounts, short[] postNeuronStateCounts, byte curState);
     public abstract boolean getConnectionIncreaseStrength(byte preNeuronState, byte postNeuronState, byte connectionType);
     public abstract byte getConnectionCreation(byte preNeuronState, byte postNeuronState);
 }
